+++ R/chromote_session.R
@@ -544,7 +544,7 @@ ChromoteSession <- R6Class(
     #'   active debugging session?
     mark_closed = function(target_closed) {
       private$session_is_active <- FALSE
-      private$target_is_active <- target_closed
+      private$target_is_active <- !target_closed
     },
 
     #' @description Retrieve active status
@@ -575,6 +575,33 @@ ChromoteSession <- R6Class(
     #' For internal use only.
     get_init_promise = function() {
       private$init_promise_
+    },
+
+    #' @description Summarise the current state of the object.
+    #' @param verbose The print method defaults to a brief summary
+    #'   of the most important debugging info; use `verbose = TRUE` tp
+    #'   see the complex R6 object.
+    #' @param ... Passed on to `format()` when `verbose` = TRUE
+    print = function(..., verbose = FALSE) {
+      if (verbose) {
+        cat(format(self, ...), sep = "\n")
+      } else {
+        if (self$is_active()) {
+          state <- "session + target active"
+        } else if (private$target_is_active) {
+          state <- "target active"
+        } else {
+          state <- "closed"
+        }
+
+        cat_line("<ChromoteSession> (", state, ")")
+        if (self$is_active())
+          cat_line("  Session ID: ", self$get_session_id())
+        if (private$target_is_active)
+          cat_line("  Target ID:  ", self$get_target_id())
+        cat_line("  Parent PID: ", self$parent$get_browser()$get_process()$get_pid())
+      }
+      invisible(self)
     },
 
     #' @field parent [`Chromote`] object